Workflow manager: consider validation batch complete if signature exists. (#1263)
I think this is safe for both new-style (single-object) & old-style
(multi-object) batches. Single-object batches are safe because they
really are just a single object, so checking for that one object is the
correct thing to do. Multi-object batches are safe because the signature
object was always written last (& only if writing the header & packet
objects succeeded), and both AWS & GCP offer strong read-after-write &
list consistency [1][2], so if the signature object was written we know
the batch & packet objects will also be written.

[1] https://docs.aws.amazon.com/AmazonS3/latest/userguide/Welcome.html#ConsistencyModel
[2] https://cloud.google.com/storage/docs/consistency