terraform: speed up prep targets in Makefile
Historically, we blow away `.terraform` (or
`cluster_bootstrap/.terraform`, depending on which `prep` target is
invoked) before initializing `terraform`. The idea is to avoid the risk
of one environment's `.tfstate` bleeding into another, which `terraform
init` can sometimes sneakily do to you.

While working on #1274, I found myself switching between remote and
local Terraform backends a lot. My implementation plans had a lot of `rm
-rf .terraform` in them, in accordance with the `prep` and
`prep-cluster-bootstrap` targets in `terraform/Makefile`. The downside
of destroying the entire directory is that you have to reinstall the
Terraform providers every time you `plan` or `apply`. I learned that
what's really important is the file `.terraform/terraform.tfstate`, and
that if we destroy just that file in the `prep` targets, we avoid the
risk of cross-contamination of backends but can speed up `prep`
considerably by not re-downloading providers.