Generate sample data deterministically from packet UUID.
The UUID is itself random, so sample data is still random; the upside of
this is that we can now determine a given sample ingestion packet's data
without reading & decrypting the data shares. This will be handy for a
validator which wants to make sure the sums we are emitting match up
with the ingestion packets included in those sums.

[Note that since even "random" v4 UUIDs contain a few deterministic
bytes, part of the data will be deterministic; I think that's OK.]