Rusoto operations: add 10-minute timeout. (#1286)
* Rusoto operations: add 10-minute timeout.

I spent a lot of today looking at this code -- in #1285, all three stack
traces are getting stuck in S3Transport::put, and more specifically in
the block_on call modified by this commit. I didn't find a smoking gun,
but I'd like to add a timeout to this request.

I suppose there are two possibilities for this deadlock:
 1) The deadlock is a true "thread" deadlock, e.g. waiting on a mutex
    that the current thread already owns. In this case, I'm not sure the
    new timeout will help.
 2) The deadlock is due to some future in create_multipart_upload not
    correctly implementing the future interface, e.g. not arranging for
    the wakeup to occur when progress can be made. In this case, the
    timeout will effectively sidestep the bug (in the sense that we'll
    now give up after waiting ten minutes, rather than waiting forever),
    and we'll have a new error in the logs confirming that this is what
    is happening.

That is, this commit may provide a workaround for the issue and even if
it doesn't, won't put us in a worse position; and either way, we have
more information. I suspect that we're in situation #2 since, from how
I'm interpreting the stack traces, we don't appear to be waiting in
application code (but async is complex enough that I might reasonably be
incorrect about that reasoning.)

Why ten minutes? That's how long retry_request is hardwired to try
before giving up, so it seemed like an OK choice.