Key rotator: add flags to skip manifest pre/post-update validations.
These are unsafe to use in general--the pre-update validations serve as
a check that the key-rotator is running against a consistent set of
(manifest, k8s secret), the post-update validations serve as a check
that our update logic didn't do something very unexpected. But they may
be necessary if the key-rotator needs to run to fix an environment that
is already out-of-sync (and therefore triggering validation failures).