Don't copy forward serialized key data if public keys don't match.
Previously, this was impossible because we validated that manifest key
data matched the secret key data before we started operating. Since we
can now bypass that check, we need to update this logic to check if keys
need to be updated even if they already have some key material.

Tests are still failing due to post-update validations that require
serialized key data to match exactly, but I think behavior is correct
now. Next commit will fixup the validations.