Add expected_sample_sum_for_batches. (#1300)
This function produces the expected sum from a SumPart comprised of the
given batch UUIDs. This turned out to be a lot simpler than I expected:
I didn't need to read the validation batches in the sum part to get
their UUIDs because the validation batch uses the same UUID as the
ingestion UUID (and the ingestion UUID is what determines the data
included in the ingestion batch, along with the "static"
dimension/packet_count parameters).

The test is a "weak" test in that it only tests that the result of
expected_sample_sum_for_batches matches up with the sum of a few
repeated calls to generate_ingestion_sample -- it does not test the full
intake -> ingest -> aggregate flow from end-to-end. The goal of this
test is to make sure these two functions agree on the sums they believe
they are generating.