Coalesce DSP global manifest versions 0 & 1.
Like the DSP locality-specific manifest versions, we don't actually need
two different message versions here -- we can define a single message
that can accept both versions. The one downside is that aws_account_id
is now always optional (previously it was required for V0 messages, and
disallowed for V1 messages) -- this means that our parsing is slightly
less strict. But we no longer use aws_account_id anywhere, so IMO it is
unlikely that this will matter.