Coalesce DSP-specific manifests versions 1 & 2 into a single version. (#1330)
The only difference between the two versions it that V2 added an
optional "peer_validation_identity" field which defaulted to None.
Generic code that attempted to extract a peer_validation_identity from a
V1 manifest would get None, as well. Since this isn't a
backwards-incompatible change, we might as well simplify things by going
back to only a single version.