Fix bug around task-deadline extension.
Previously, after the first time we extended a task's deadline, we would
attempt to extend the deadline again every time we checked if the
deadline should be updated. Now, every time we extend a task's deadline,
we "reset" the starting point so that we only attempt to reset the
deadline again once enough time has passed.