Fix bug around task-deadline extension. (#1333)
Previously, after the first time we extended a task's deadline, we would
attempt to extend the deadline again every time we checked if the
deadline should be updated. Now, every time we extend a task's deadline,
we reset the last-refresh time so that we only attempt to reset the
deadline again once enough time has passed.