Check in .terraform.lock.hcl files
When terraform is run for the first time, it creates these files and
recommends including them in version control. Checking them in means it
will use the exact same provider versions on each computer, and on each
run of `terraform init`. Dependabot advertises that it supports
Terraform lockfiles, so presumably they will be updated in provider
update PRs.