facilitator: correctly report included batch UUIDs
If an ingestion batch contains a packet that cannot be decrypted, then
the intake task for it will fail, currently producing a dead letter.
However, we gracefully handle this case at aggregation time, because if
one batch out of `n` contains an undecryptable packet, we still want to
aggregate over the other `n - 1` batches. However, when discarding a
batch from the aggregate, we were still including its batch UUID in the
sum part message sent to the portal server, which causes our
aggregator's list of batch UUIDs to not match up with the peer's, which
the portal server rightfully considers an error.

This commit makes the aggregator only include a batch UUID if the
batch's packets were aggregated into the sum part.

Resolves #1414