Fix failure rate alerting, new alert for rejecting (#1420)
This fixes an issue with the expressions used for failure rate alerting.
It also introduces a new value of "rejected" for the label "status", on
the "tasks_finished" metrics, for tasks that result in a non-retryable
error. We apply the "rejected" value when we divert the task to a
separate rejected queue, rather than let it reach the dead letter queue.
A new alert is introduced to monitor the fraction of tasks being
rejected, and alert if it is too high.