During aggregation, require ingestion header parameters to match. (#1421)
Previously, we didn't check, and would happily aggregate even if some of
the parameters didn't match--but at a higher level, data with different
parameters can't be meaningfully aggreagated, so this produced
non-useful output. All we can do is notice & complain, but we should do
so rather than producing bad output.