terraform: configure log retention period
This change will require importing the log group that EKS already
created, to keep Terraform from trying to create it again and finding
that its name is already in use. For example:

Run at least one makefile target to make sure that
cluster_bootstrap/.terraform/terraform.tfstate points to the correct
backend bucket for this environment.
$ ENV=<environment name> make plan-cluster-bootstrap
Import the log group resource. (The easiest way to get the log group's
name is to attempt to apply once, and look at the log group name in the
error message)
$ ENV=<environment name> terraform -chdir=cluster_bootstrap import -var-file=../variables/<environment name>.tfvars -input=false module.eks[0].aws_cloudwatch_log_group.cluster_logs /aws/eks/prio-<environment name>/cluster
Finally, apply the changes. This should add a prio-env tag to the log
group, and set its retention period.
$ ENV=<environment name> make apply-cluster-bootstrap