terraform: Upgrade aws-for-fluent-bit container
The fluent-bit DaemonSet was previously running an old version of this
image, which included Fluent Bit 1.6. That version did not support the
`log_retention_days` configuration parameter for the `cloudwatch_logs`
plugin, and thus it would fail to start. Upgrading the container brings
the Fluent Bit version to 1.8, which does support this parameter.