facilitator: fail intake of empty ingestion batch
In prod-intl, we observed some cases where intake tasks were failing
because of empty ingestion batches causing `facilitator` to make
malformed S3 requests. We will fix the S3 client, but we can also detect
empty ingestion batches before we get around to making any requests to a
transport. We mark this as a non-retryable error since there's no point
redelivering the task for such a problem.