facilitator: cancel S3 uploads with no parts
If we call `S3Transport::put` to get a `TransportWriter`, but then call
`MultipartUploadWriter::complete_upload` without ever writing anything
to it, then we get an error from S3, which rejects our
`CompleteMultipartUpload` request that contains no `Parts` array. We now
detect that there are no complete parts and cancel the multipart upload
in S3 without reporting an error, bringing us in line with the
established behavior of the GCS `StreamingUploadWriter`. Both
implementations will now cancel the upload on detecting this condition
to avoid leaving dangling uploads in the cloud service.

The downside of this approach is that if this error occurs, then the
caller is left with a `MultipartUploadWriter` that can no longer be
used, because the upload that it refers to has been cancelled. The
proper solution here would be for `TransportWriter::complete_upload` to
take `self` instead of `&mut self`, but that is a bigger change than is
called for in this bug fix. I settled for documenting this behavior in
`TransportWriter::complete_upload` and `TransportWriter::cancel_upload`.