Spot VM revamp (#1703)
* terraform: rewrite usage of spot VMs

This creates a separate GKE node pool with spot VMs, and steers
different workloads to spot or non-spot nodes. Kubernetes system
services, metrics infrastructure, and CronJobs will be scheduled on
non-spot nodes, while facilitator deployments may be scheduled on spot
nodes. The spot nodes are assigned a NoSchedule taint, and facilitator
deployments are given a toleration for that taint, plus a node affinity
for spot VM/spot instance labels.

* Switch spot node pool from google-beta to google

Spot VMs are now generally available, as of the latest provider version.

* Namespaced taint keys