terraform: add option to collect heap profiles
This adds an option to set up an in-cluster NFS server, backed by a GCE
Persistent Disk, and has workflow-manager and key-rotator jobs write out
heap profiles to a volume backed by the NFS server. An NFS server is
necessary because we need a ReadWriteMany persistent volume, and this is
the simplest way to achieve that on GKE. (GCP has a managed NFS
offering, but its minimum filesystem size is 1 TiB)

It is intended that this new option will be turned on for a brief time
to collect profiles, the profiles will be copied out of the nfs-server
pod for later analysis, and the option will be turned off again to tear
down NFS-related infrastructure. The added NFS mounts are additional
potential failure points for these jobs, so we may as well remove them
in normal operation.