terraform: update CRD specification for k8s 1.22 (#2281)
We're migrating our EKS clusters to Kubernetes 1.22 (#2249). That
release of Kubernetes promoted the `CustomResourceDefinition` resource
from `apiextensions.k8s.io/v1beta1` to plain `v1`. That also changes the
schema for CRDs so that they have a `versions` array instead of a single
`version` ([k8s docs][1]). We update our declaration of the
`kubernetes_manifest` resource to match.

[1]: https://kubernetes.io/docs/reference/kubernetes-api/extend-resources/custom-resource-definition-v1/#CustomResourceDefinitionSpec