Update EKS plugins for 1.23 (#2499)
* Updated EKS to 1.23 in staging environments

AWS is deprecating EKS 1.22 in early June, but we need to continue
running the international instance of ENPA through the end of that
month, so we update to 1.23.

* Update EKS plugins for 1.23

This updates the EBS and VPC add-ons and the cluster autoscaler to
recent ones compatible with 1.23. Because we are only doing a single
minor version upgrade of the VPC CNI plugin, we don't need to perform
any manual updates.